Move transport and swarm builders into separate modules
