+++ aquadoggo/src/network/mod.rs
@@ -4,6 +4,8 @@ mod behaviour;
 mod config;
 mod identity;
 mod service;
+mod swarm;
+mod transport;
 
 pub use config::NetworkConfiguration;
 pub use service::network_service;
